+++ src/main/java/digit/acdemy/tutorial/repository/querybuilder/QueryBuilder.java
@@ -1,0 +1,5 @@
+package digit.acdemy.tutorial.repository.querybuilder;
+
+public interface QueryBuilder<T, V> {
+    V build(T criteria);
+}
