+++ src/main/java/digit/acdemy/tutorial/repository/querybuilder/QueryPair.java
@@ -1,0 +1,3 @@
+package digit.acdemy.tutorial.repository.querybuilder;
+
+public record QueryPair<K, V>(K query, V params) {}
