+++ src/main/java/digit/acdemy/tutorial/web/models/AdvocateRequest.java
@@ -1,0 +1,42 @@
+package digit.acdemy.tutorial.web.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.egov.common.contract.request.RequestInfo;
+import org.springframework.validation.annotation.Validated;
+import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import lombok.Data;
+import lombok.Builder;
+
+/**
+ * AdvocateRequest
+ */
+@Validated
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-01-19T00:25:41.754980941+05:30[Asia/Kolkata]")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class AdvocateRequest {
+
+    @JsonProperty("requestInfo")
+    @Valid
+    private RequestInfo requestInfo = null;
+
+    @JsonProperty("advocates")
+    @Valid
+    private List<Advocate> advocates = null;
+
+    public AdvocateRequest addAdvocatesItem(Advocate advocatesItem) {
+        if (this.advocates == null) {
+            this.advocates = new ArrayList<>();
+        }
+        this.advocates.add(advocatesItem);
+        return this;
+    }
+}
