+++ src/main/java/digit/acdemy/tutorial/web/models/AdvocateSearchCriteria.java
@@ -1,0 +1,29 @@
+package digit.acdemy.tutorial.web.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.springframework.validation.annotation.Validated;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import lombok.Data;
+import lombok.Builder;
+
+/**
+ * AdvocateSearchCriteria
+ */
+@Validated
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-01-19T00:25:41.754980941+05:30[Asia/Kolkata]")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class AdvocateSearchCriteria {
+
+    @JsonProperty("id")
+    private String id = null;
+
+    @JsonProperty("barRegistrationNumber")
+    private String barRegistrationNumber = null;
+
+    @JsonProperty("applicationNumber")
+    private String applicationNumber = null;
+}
