--- src/main/java/digit/config/Configuration.java
@@ -1,85 +1,0 @@
-package digit.config;
-
-import lombok.*;
-import org.egov.tracer.config.TracerConfiguration;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Import;
-import org.springframework.stereotype.Component;
-
-@Component
-@Data
-@Import({TracerConfiguration.class})
-@NoArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-public class Configuration {
-
-
-    // User Config
-    @Value("${egov.user.host}")
-    private String userHost;
-
-    @Value("${egov.user.context.path}")
-    private String userContextPath;
-
-    @Value("${egov.user.create.path}")
-    private String userCreateEndpoint;
-
-    @Value("${egov.user.search.path}")
-    private String userSearchEndpoint;
-
-    @Value("${egov.user.update.path}")
-    private String userUpdateEndpoint;
-
-
-    //Idgen Config
-    @Value("${egov.idgen.host}")
-    private String idGenHost;
-
-    @Value("${egov.idgen.path}")
-    private String idGenPath;
-
-
-    //Workflow Config
-    @Value("${egov.workflow.host}")
-    private String wfHost;
-
-    @Value("${egov.workflow.transition.path}")
-    private String wfTransitionPath;
-
-    @Value("${egov.workflow.businessservice.search.path}")
-    private String wfBusinessServiceSearchPath;
-
-    @Value("${egov.workflow.processinstance.search.path}")
-    private String wfProcessInstanceSearchPath;
-
-
-    //MDMS
-    @Value("${egov.mdms.host}")
-    private String mdmsHost;
-
-    @Value("${egov.mdms.search.endpoint}")
-    private String mdmsEndPoint;
-
-
-    //HRMS
-    @Value("${egov.hrms.host}")
-    private String hrmsHost;
-
-    @Value("${egov.hrms.search.endpoint}")
-    private String hrmsEndPoint;
-
-
-    //URLShortening
-    @Value("${egov.url.shortner.host}")
-    private String urlShortnerHost;
-
-    @Value("${egov.url.shortner.endpoint}")
-    private String urlShortnerEndpoint;
-
-
-    //SMSNotification
-    @Value("${egov.sms.notification.topic}")
-    private String smsNotificationTopic;
-}
