--- src/main/java/digit/kafka/Consumer.java
@@ -1,21 +1,0 @@
-package digit.kafka;
-
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.stereotype.Component;
-
-import java.util.HashMap;
-
-@Component
-public class Consumer {
-
-    /*
-    * Uncomment the below line to start consuming record from kafka.topics.consumer
-    * Value of the variable kafka.topics.consumer should be overwritten in application.properties
-    */
-    //@KafkaListener(topics = {"kafka.topics.consumer"})
-    public void listen(final HashMap<String, Object> record) {
-
-        //TODO
-
-    }
-}
