--- src/main/java/digit/kafka/Producer.java
@@ -1,20 +1,0 @@
-package digit.kafka;
-
-import lombok.extern.slf4j.Slf4j;
-import org.egov.tracer.kafka.CustomKafkaTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-// NOTE: If tracer is disabled change CustomKafkaTemplate to KafkaTemplate in autowiring
-
-@Service
-@Slf4j
-public class Producer {
-
-    @Autowired
-    private CustomKafkaTemplate<String, Object> kafkaTemplate;
-
-    public void push(String topic, Object value) {
-        kafkaTemplate.send(topic, value);
-    }
-}
